refactor matmul files
