+++ crates/burn-jit/src/kernel/matmul/mod.rs
@@ -1,13 +1,23 @@
 mod base;
-mod mem_coalescing;
-mod tiling2d;
+mod simple;
 mod tune;
 
 /// Contains utilitary for matmul operation
 pub mod utils;
 
 pub use base::*;
-pub use mem_coalescing::*;
-pub use tiling2d::*;
+pub use simple::*;
 pub use tune::*;
 pub use utils::*;
+
+#[cfg(feature = "export_tests")]
+#[allow(missing_docs)]
+pub mod padding;
+
+#[cfg(not(feature = "export_tests"))]
+mod padding;
+
+pub mod tiling2d;
+pub mod tiling2d_padded;
+pub use tiling2d::*;
+pub use tiling2d_padded::*;
